+++ packages/terra-dropdown-button/src/DropdownButton.jsx
@@ -65,8 +65,8 @@ class DropdownButton extends React.Component {
     this.setState(prevState => ({ isOpen: !prevState.isOpen }));
   }
 
-  handleDropdownRequestClose() {
-    this.setState({ isOpen: false });
+  handleDropdownRequestClose(callback) {
+    this.setState({ isOpen: false }, typeof callback === 'function' ? callback : undefined);
   }
 
   /*