+++ packages/terra-dropdown-button/src/_DropdownList.jsx
@@ -107,8 +107,7 @@ class DropdownList extends React.Component {
       event.preventDefault();
       if (this.pressed) {
         const item = Util.findByIndex(this, this.state.focused);
-        this.props.requestClose();
-        item.props.onSelect(event, item.props.metaData);
+        this.props.requestClose(item.props.onSelect(event, item.props.metaData));
       }
       this.pressed = false;
     }